# swagger/swaggerfile.json
{
    "info": {
        "version": "0.0.1", 
        "title": "CrowdSauce"
    }, 
    "paths": {
...
        "/api/tags/": {
            "get": {
                "operationId": "add a tag",
                "tags": [ "tags" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": false,
                        "type": "integer",
                        "description": "UserID of user doing the favoriting",
                        "in": "header",
                        "name": "userid"
                    },
                    {
                        "required": false,
                        "type": "string",
                        "description": "The facebook access token of the user",
                        "in": "header",
                        "name": "accesstoken"
                    }
                ],
                "description": "Favorite a post"
            },
            "put": {
                "operationId": "add a tag",
                "tags": [ "tags" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true,
                        "type": "integer",
                        "description": "UserID of user doing the favoriting",
                        "in": "header",
                        "name": "userid"
                    },
                    {
                        "required": false,
                        "type": "string",
                        "description": "The facebook access token of the user",
                        "in": "header",
                        "name": "accesstoken"
                    },
                    {
                        "required": true,
                        "description": "Contains data about the tag to be posted",
                        "in": "body",
                        "name": "body",
                        "schema": {
                            "$ref": "#/definitions/Tag"
                        }
                    }
                ],
                "description": "Favorite a post"
            },
            "post": {
                "operationId": "add a tag",
                "tags": [ "tags" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true,
                        "type": "integer",
                        "description": "UserID of user doing the favoriting",
                        "in": "header",
                        "name": "userid"
                    },
                    {
                        "required": false,
                        "type": "string",
                        "description": "The facebook access token of the user",
                        "in": "header",
                        "name": "accesstoken"
                    },
                    {
                        "required": true,
                        "description": "Contains data about the tag to be posted",
                        "in": "body",
                        "name": "body",
                        "schema": {
                            "$ref": "#/definitions/Tag"
                        }
                    }
                ],
                "description": "Favorite a post"
            },
            "delete": {
                "operationId": "delete tag",
                "tags": [ "tags" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true,
                        "type": "integer",
                        "description": "UserID for user doing the unfavoriting",
                        "in": "header",
                        "name": "userid"
                    },
                    {
                        "required": false,
                        "type": "string",
                        "description": "The facebook access token of the user",
                        "in": "header",
                        "name": "accesstoken"
                    },
                    {
                        "required": true,
                        "description": "Contains data about the tag to be deleted",
                        "in": "body",
                        "name": "body",
                        "schema": {
                            "$ref": "#/definitions/Tag"
                        }
                    }
                ],
                "description": "Unfavorite a post"
            }
        },
...
        "/api/tags/tag/": {
            "get": {
                "operationId": "get post favorites",
                "tags": [ "tags" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {
                        "description": "Success",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/FavoriteReturn"
                            },
                            "type": "array"
                        }
                    },
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": false,
                        "type": "integer",
                        "description": "UserID for user",
                        "in": "header",
                        "name": "userid"
                    },
                    {
                        "required": false,
                        "type": "string",
                        "description": "The facebook access token of the user",
                        "in": "header",
                        "name": "accesstoken"
                    },
                    {
                        "required": true,
                        "type": "string",
                        "description": "The post to get a favorites for",
                        "in": "query",
                        "name": "postId"
                    }
                ],
                "description": "Get all posts for a given user"
            }
        },
...
    }, 
    "schemes": [
        "http", 
        "https"
    ], 
    "produces": [
        "application/json"
    ], 
    "basePath": "/", 
    "host": "localhost:3000",
    "definitions": {
...
        "Tag": {
            "properties": {
                "postId": {
                    "tagName": "string"
                }
            }
        },
...
        "FavoriteReturn": {
            "properties": {
                "postId": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        }
    }, 
    "swagger": "2.0", 
    "consumes": [
        "application/json"
    ]
}